Create assets change notifier
