+++ il_core/lib/il_core.dart
@@ -2,6 +2,7 @@ library core;
 
 export 'src/logic/auth/login_handler.dart';
 export 'src/logic/display_assets/asset_display_handler.dart';
+export 'src/logic/display_assets/assets_change_notifier.dart';
 export 'src/logic/reports/asset_report_generator.dart';
 
 export 'src/context/backend_context.dart';
